[#3016] Most of the REST API documentation.
There are examples for pretty much everything that can be accessed via
the rest API.  There's still plenty todo to make this better, but we
should get something up for now.

I'll be getting this reviewed by another developer tomorrow as long as
no urgent bugs get in the way.

TODO: (I'll create separate tickest for these)

Finish documenting the security file settings.  I documented the
security for the restful interface, but not the file specific security
settings.

List the supported formats for each API call.  Currently this is
difficult and many formats that are supposedly supported return bad data
like the stringification of a class.

Figure out why report only allows PUTS not GETS, and PUTS fail with a (0
of 1) argument error.

The Key at the end of the URI should be optional if it's not necessary.

Sample return text for the curl calls